tests/wpt/web-platform-tests/app-history/navigate/return-value/navigate-interrupted.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  const result1 = appHistory.navigate("#1");
  const result2 = appHistory.navigate("#2");

  assert_equals(appHistory.entries().length, 3);
  assert_array_equals(appHistory.entries().map(e => (new URL(e.url)).hash), ["", "#1", "#2"]);

  await assertCommittedFulfillsFinishedRejectsDOM(t, result1, appHistory.entries()[1], "AbortError");
  await assertBothFulfill(t, result2, appHistory.current);
}, "interrupted navigate() promises");
</script>
